W-113863887 (refactor): split classes in SpecSettings file into their own files
