+++ amf-api-contract/shared/src/main/scala/amf/apicontract/internal/spec/async/parser/context/Async20Syntax.scala
@@ -1,6 +1,7 @@
 package amf.apicontract.internal.spec.async.parser.context
 
-import amf.shapes.internal.spec.common.parser.{Async20ShapeSyntax, SpecSyntax}
+import amf.shapes.internal.spec.async.parser.Async20ShapeSyntax
+import amf.shapes.internal.spec.common.parser.SpecSyntax
 
 object Async20Syntax extends SpecSyntax {
   override val nodes: Map[String, Set[String]] = Async20ShapeSyntax.nodes ++ Map(
