+++ amf-api-contract/shared/src/main/scala/amf/apicontract/internal/spec/async/parser/context/Async20WebApiContext.scala
@@ -4,7 +4,10 @@ import amf.apicontract.internal.spec.common.AsyncWebApiDeclarations
 import amf.core.client.scala.config.ParsingOptions
 import amf.core.client.scala.parse.document.{ParsedReference, ParserContext}
 import amf.core.internal.remote.{AsyncApi20, Spec}
-import amf.shapes.internal.spec.common.parser.{Async2Settings, SpecSyntax}
+import amf.shapes.internal.spec.async
+import amf.shapes.internal.spec.async.parser.Async2Settings
+import amf.shapes.internal.spec.common.parser.SpecSyntax
+import amf.shapes.internal.spec.oas.parser
 
 import scala.collection.mutable
 
@@ -15,7 +18,15 @@ class Async20WebApiContext(
     private val ds: Option[AsyncWebApiDeclarations] = None,
     private val operationIds: mutable.Set[String] = mutable.HashSet(),
     options: ParsingOptions = ParsingOptions()
-) extends AsyncWebApiContext(loc, refs, options, wrapped, ds, operationIds, Async2Settings(Async20Syntax)) {
+) extends AsyncWebApiContext(
+      loc,
+      refs,
+      options,
+      wrapped,
+      ds,
+      operationIds,
+      async.parser.Async2Settings(Async20Syntax)
+    ) {
   override val factory: Async20VersionFactory = Async20VersionFactory()(this)
 
   override def makeCopy(): Async20WebApiContext =
