+++ amf-api-contract/shared/src/main/scala/amf/apicontract/internal/spec/oas/parser/context/Oas2WebApiContext.scala
@@ -4,8 +4,10 @@ import amf.apicontract.internal.spec.common.OasWebApiDeclarations
 import amf.core.client.scala.config.ParsingOptions
 import amf.core.client.scala.parse.document.{ParsedReference, ParserContext}
 import amf.core.internal.remote.{Oas20, Spec}
-import amf.shapes.internal.spec.common.parser.{Oas2Settings, SpecSyntax}
+import amf.shapes.internal.spec.common.parser.SpecSyntax
 import amf.shapes.internal.spec.common.{OAS20SchemaVersion, SchemaPosition, SchemaVersion}
+import amf.shapes.internal.spec.oas.parser
+import amf.shapes.internal.spec.oas.parser.Oas2Settings
 
 class Oas2WebApiContext(
     loc: String,
@@ -13,7 +15,7 @@ class Oas2WebApiContext(
     wrapped: ParserContext,
     ds: Option[OasWebApiDeclarations] = None,
     options: ParsingOptions = ParsingOptions()
-) extends OasWebApiContext(loc, refs, options, wrapped, ds, Oas2Settings(Oas2Syntax)) {
+) extends OasWebApiContext(loc, refs, options, wrapped, ds, parser.Oas2Settings(Oas2Syntax)) {
   override val factory: Oas2VersionFactory = Oas2VersionFactory()(this)
 
   override val defaultSchemaVersion: SchemaVersion = OAS20SchemaVersion.apply(SchemaPosition.Other)
