+++ amf-api-contract/shared/src/main/scala/amf/apicontract/internal/spec/oas/parser/context/Oas3WebApiContext.scala
@@ -3,8 +3,9 @@ package amf.apicontract.internal.spec.oas.parser.context
 import amf.apicontract.internal.spec.common.OasWebApiDeclarations
 import amf.core.client.scala.config.ParsingOptions
 import amf.core.client.scala.parse.document.{ParsedReference, ParserContext}
-import amf.shapes.internal.spec.common.parser.Oas3Settings
 import amf.shapes.internal.spec.common.{OAS30SchemaVersion, SchemaPosition, SchemaVersion}
+import amf.shapes.internal.spec.oas.parser
+import amf.shapes.internal.spec.oas.parser.Oas3Settings
 
 class Oas3WebApiContext(
     loc: String,
@@ -12,7 +13,7 @@ class Oas3WebApiContext(
     private val wrapped: ParserContext,
     private val ds: Option[OasWebApiDeclarations] = None,
     options: ParsingOptions = ParsingOptions()
-) extends OasWebApiContext(loc, refs, options, wrapped, ds, Oas3Settings(Oas3Syntax)) {
+) extends OasWebApiContext(loc, refs, options, wrapped, ds, parser.Oas3Settings(Oas3Syntax)) {
   override val factory: Oas3VersionFactory = Oas3VersionFactory()(this)
 
   override val defaultSchemaVersion: SchemaVersion = OAS30SchemaVersion.apply(SchemaPosition.Other)
