+++ amf-api-contract/shared/src/main/scala/amf/apicontract/internal/spec/oas/parser/context/OasSyntax.scala
@@ -1,6 +1,7 @@
 package amf.apicontract.internal.spec.oas.parser.context
 
-import amf.shapes.internal.spec.common.parser.{Oas2ShapeSyntax, Oas3ShapeSyntax, SpecSyntax}
+import amf.shapes.internal.spec.common.parser.SpecSyntax
+import amf.shapes.internal.spec.oas.parser.{Oas2ShapeSyntax, Oas3ShapeSyntax}
 
 object Oas3Syntax extends SpecSyntax {
   override val nodes: Map[String, Set[String]] = Oas3ShapeSyntax.nodes ++ Map(
