+++ amf-shapes/shared/src/main/scala/amf/shapes/internal/spec/oas/parser/OasLikeIgnoreCriteria.scala
@@ -1,0 +1,13 @@
+package amf.shapes.internal.spec.oas.parser
+
+import amf.shapes.internal.spec.common.parser.IgnoreCriteria
+
+object OasLikeIgnoreCriteria extends IgnoreCriteria {
+
+  private val shapesThatDontPermitRef = List("paths", "operation")
+
+  override def shouldIgnore(shape: String, property: String): Boolean = {
+    property.startsWith("x-") || (property == "$ref" && !shapesThatDontPermitRef.contains(shape)) || (property
+      .startsWith("/") && shape == "paths")
+  }
+}
